MS Custom Connectors - Reassigning of any approval/cancellation
Here we have created custom connectors that we can use for Reassigning of any approval/cancellation and other action for those workflows that is already running and waiting for some Person. But if the person is not available to approve or if we want to reassign in between we can create custom connectors and follow steps provided.